+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - New `metadata` condition type.
+- More metadata fields for `kafka` input.
 
 ### 0.23.1 - 2018-08-06